migrate assing tests
